+++ source/MaterialXGenEssl/EsslSyntax.cpp
@@ -1,0 +1,20 @@
+//
+// TM & (c) 2021 Lucasfilm Entertainment Company Ltd. and Lucasfilm Ltd.
+// All rights reserved.  See LICENSE.txt for license.
+//
+
+#include <MaterialXGenEssl/EsslSyntax.h>
+
+#include <MaterialXGenShader/Shader.h>
+#include <MaterialXGenShader/ShaderGenerator.h>
+
+#include <memory>
+
+namespace MaterialX
+{
+
+EsslSyntax::EsslSyntax()
+{
+}
+
+}